In the HelloWorldServlet, I changed the urlPatterns so that my CSS would work.
In the CSS file, I did many things, but have so far commented everything out to get the NavBar the way we wanted, and after this commit I will continue working on it to get it up to speed.

In the head file, I commented out my link to the stylesheet, added the jQuery and bootstrap CDN script tags so that the dropdown in the NavBar would function correctly with the bootstrap framework we have implemented.

In the navbar file, I went through many changes, and have finally got the code to work so that the NavBar looks the way it should, and the dropdown functions now that we added the necessary script tags.

Deleted ListAdsDao file to prevent merge errors with the group - since Brad deleted it on his work.